refactor: PPS-41 major refactoring of HTTP classes. Extracted API inside an api package. Started working on a small DSL-like http request API
